Spin 2.12 distinctBy adapter out into its own module
