+++ modules/codegen/src/main/scala-2.13/com/twilio/guardrail/generators/syntax/Specialized.scala
@@ -1,0 +1,3 @@
+package com.twilio.guardrail.generators.syntax
+
+trait SpecializedSyntax
