+++ modules/codegen/src/main/scala/com/twilio/guardrail/generators/syntax/package.scala
@@ -6,9 +6,6 @@ import java.util.regex.Matcher.quoteReplacement
 import io.swagger.v3.oas.models.{ Operation, PathItem }
 import io.swagger.v3.oas.models.media.Schema
 import io.swagger.v3.oas.models.parameters.Parameter
-import scala.collection.IterableLike
-import scala.collection.generic.CanBuildFrom
-import scala.language.implicitConversions
 
 package syntax {
   class RichNotNullShower[A](value: A) {
@@ -16,27 +13,9 @@ package syntax {
     def showNotNullIndented(indent: Int): String =
       ("  " * indent) + value.toString().linesIterator.filterNot(_.contains(": null")).mkString("\n" + ("  " * indent))
   }
-
-  class RichCollection[A, Repr](xs: IterableLike[A, Repr]) {
-    @SuppressWarnings(Array("org.wartremover.warts.Var"))
-    def distinctBy[B, That](f: A => B)(implicit cbf: CanBuildFrom[Repr, A, That]): That = {
-      val builder = cbf(xs.repr)
-      val i       = xs.iterator
-      var set     = Set[B]()
-      while (i.hasNext) {
-        val o = i.next
-        val b = f(o)
-        if (!set(b)) {
-          set += b
-          builder += o
-        }
-      }
-      builder.result
-    }
-  }
 }
 
-package object syntax {
+package object syntax extends SpecializedSyntax {
   val GENERATED_CODE_COMMENT_LINES: List[String] = List(
     "This file was generated by guardrail (https://github.com/twilio/guardrail).",
     "Modifications will be overwritten; instead edit the OpenAPI/Swagger spec file."
@@ -102,6 +81,4 @@ package object syntax {
   implicit def RichOperation: Operation => RichNotNullShower[Operation] = new RichNotNullShower[Operation](_)
   implicit def RichPathItem: PathItem => RichNotNullShower[PathItem]    = new RichNotNullShower[PathItem](_)
   implicit def RichParameter: Parameter => RichNotNullShower[Parameter] = new RichNotNullShower[Parameter](_)
-
-  implicit def toRichCollection[A, Repr](xs: IterableLike[A, Repr]): RichCollection[A, Repr] = new RichCollection(xs)
 }
